requirements.txt
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
streamlit>=1.28.0
requests>=2.25.0